requirements.txt
Django==2.0.7
django-tastypie==0.14.1
pkg-resources==0.0.0
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.5
six==1.11.0
